add a test for comparing Q{Object,Variant} wrappers from script
In the old back-end, this worked because we extended the ECMA
comparison algorithm to treat QObject and QVariant wrappers
specially. Instead of comparing the script objects (which would
always fail), we compared the wrapped values (which could be the
same in both wrappers).
In the new JSC-based back-end this currently fails. We would have
to add a hack (QtScript-specific ifdef) to JSC in order to support
it.