Added a behaviour parameter to QXmlStreamReader::readElementText
This makes the function a bit more useful, since previously it was only
safe to use if you were sure that it would not encounter an unexpected
child element, or if you would be alright with canceling the parser on
such an occurrence.

Now it is also possible to have it ignore any unexpected child elements,
or to have it include the text found in any child elements.

Task-number: 231938
Reviewed-by: mae