Dont create native windows when setting up gestures.
The code used to create native window handle even before checking if the widget
subscribes to gestures, basically breaking alien widgets. Check the position of
the gesture instead when the WM_GESTURE message is received in the GID_BEGIN
state. This will work properly in most cases - but not always as the position
that we get with the WM_GESTURE message is the center point of the multitouch
gesture, which might be outside of the widget depending on positions of
fingers.

Reviewed-by: trustme