Add projection support to QMatrix4x4::toTransform()
When converting from 3D back to 2D, it is necessary to project
the z component unless an orthographic projection is desired.
This change adds a distanceToPlane argument that specifies the
perspective projection factor to apply during the conversion.
The default value of 1024 corresponds to the projection performed
by QTransform::rotate(), for consistency with existing classes.

Reviewed-by: trustme