Autotest: add one more check for operator==
I'm running out of ideas as to why MSVC.NET 2003 cannot do this test:
    baseptr == aData

If I do baseptr.data() == aData, it works. And the operator== that it
should be calling does exactly that. So my only clue so far is that it's
calling some other operator== -- which doesn't make sense, since there
is no other.