Started fleshing out the VideoPlayer implementation
Now loads, prepares and plays a clip, but the video is not visible because it's not yet wired up to a VideoWidget.

Video 'playback' can be tested using the demos/mediaplayer application, but the menus are not displayed properly, so a video clip filename must be hardcoded in main.cpp and passed to the MediaPlayer constructor.