Handle Touch* events in the same way as other input events in QAbstractScrollArea
We send the Touch* events filtered from the viewport to the scroll
area's default handler so that we can take advantage of the default
handling in QWidget. But we also need to ignore Touch* events in
QAbstractScrollArea::event() so that we don't get the events twice
(once from the filter, and again if propagated).

Reviewed-by: Richard Moe Gustavsen