Fix linking to uninstalled libraries after LIBS_PRIVATE.
On ELF systems, the static linker isn't allowed to find a linked
library's dependencies using the -L flag. That means if you're linking
lib or app X against a library A, and library A links against library
B, then -L$QTDIR/lib won't apply to B.

Before LIBS_PRIVATE, we had -lB, so it did apply. Now we need to find
another solution.

The solution is one of:
 a) install the libraries before linking anything against them
 b) set LD_LIBRARY_PATH
 c) set -rpath or -rpath-link

Reviewed-by: TrustMe