Revert "Remove a global destructor from QtDBus."
This reverts commit 3ed5b5ebf5b360dedbba14c03f6ca5701b3b9290. That was
a well-intentioned commit and technically correct, but it breaks when
there are static QDBusConnection objects in something that is loaded
before QtDBus (like libraries that depend on QtDBus or applications
like qdbus)

Since QDBusConnection is ref-counted, the last instance is supposed to
disconnect. However, if we have already unloaded libdbus-1, we can't
do that. Q_GLOBAL_STATICs are destroyed too soon.