Use glXGetProcAddress to resolve glx extensions
On systems where the GL driver is pluggable (like Mesa), we have to use
the glXGetProcAddressARB extension to resolve other function pointers as
the symbols wont be in the GL library, but rather in a plugin loaded by
the GL library.

This fix basically makes texture-from-pixmap work on Mesa drivers like
intel i915 & friends.

Reviewed-by: Trond