Gestures, Mac: remove pan gesture from qabstractscrollarea
On Mac there really is no native pan gesture. Instead, this
gesture is delivered as mouse wheel events. So adding a
pan gesture to qabstractscrollarea means that it will get
both wheel and pan events. And both will try to move the
viweport. I think that the correct thing to do here is
to not apply a pan gesture by default on all scroll areas.