Don't float objects into next page unless they can fit in it
If you had a floating table which spanned over several pages, this would
always float down to the next page, even when it was the only element
in the document. The main reason why this mechanism was added to
positionFloat() was to handle floating images, since they can't be
broken down into more than one page. Tables, however, have logic to
ensure they are broken into several pages if they are too large for the
current page. There is however no reason to float an object into the
next page if it is larger than that page.

Since the size of tables are altered by their position as long as they
span more than one page, we need to set the sizeDirty flag for
floating tables when positioning and then relayout the frame to make
sure the page breaks are set correctly.

Task-number: 240086
Reviewed-by: Simon Hausmann