Fix Qt::ClickFocus for QGraphicsProxyWidget.
This fixes click-focus for the Embedded Dialogs demo and the Pad Navigator
demo. In Qt 4.6, the behavior of the spontaneous-bit in events has
changed. This has happened in a number of submits that also modify
Graphics View to have the correct behavior. Somewhere on the line,
the spontaneous bit has been reset, preventing QApplicationPrivate's
ClickFocus handling to kick in. This again causes the Embedded Dialogs
demo to not give focus to widgets that are clicked.

This fix is not likely to be 100% correct. What it does is ensure that
the spont-bit survives a call to qt_sendSpontaneousEvent (as this bit
is for some reason reset at the end of QApplicationPrivate::notify_helper),
and this fixes the most apparent regression in ClickFocus handling.
Pending closer review.

The fix includes an autotest that should be kept if the fix is reverted.

Reviewed-by: bnilsen