Mac: When using Qt as a plugin app, menubars does not work
It seems like we don't respect the Qt::AA_MacPluginApplication
attribute. Setting this attribute means that the native application
should control the menu bar. This patch does a check for this, and
leaves the menubar alone if its set.
We could consider allowing menubars if the Qt::AA_DontUseNativeMenuBar
is set (those should placed inside the window)

Reviewed-by: msorvig