Implement IME reconversion on windows.
Windows IME supports reconversion of text. e.g. On a Japanese layout,
up on pressing the HENKAN key a list of choices for the current word are
shown in a popup. This patch adds that support to Qt. We will select the
current word in the widget and the choices are shown as in the editing
mode.

Task-number:225588
Reviewed-by: axis