Made video playback visible using a couple of hacks
These hacks, which are enabled using macros in defs.h, are:

PHONON_MMF_HARD_CODE_VIDEO_RECT
VideoPlayer should obtain the screen rectangle into which video rendering will occur, using the CoeControl associated with the video widget (i.e. the VideoOutput object).  However, this control always has co-ordinates (relative to its parent) of (0,0)-(100,30), regardless of whether updateGeometry() and show() are called on the widget.  So, this macro just hard-codes the screen rectangle to a value which works when the breakfast.mp4 clip is used for testing.

PHONON_MMF_EXPLICITLY_SHOW_VIDEO_WIDGET
In order that the video region does not overwrite the slider and buttons below it, the parent widget of the VideoObject instance must be appropriately re-sized.  Debugging showed that both the parent and grandparent of VideoObject are in fact invisible at the time playback begins, so for now we just manually call show() on the grandparent.

Clearly both of these are only temporary measures, for use until the underlying cause of the problem has been determined.