Cocoa, the menus can be disabled after a modal dialog
Why this happends is a bit blurry. From before, I know that
cocoa is a bit buggy regarding setting a menu item
hidden or not. The solution back then resulted in the function
syncNSMenuItemEnabled in qmenu_mac.mm. This patch basically
applies the same (silly) trick; disabling the menuitem before
enabling it. This seems to force an update to the menu items
enabled state. For the record: this is not a fix that I
embrace. I hope we can remove it again some day. See
task for how to reproduce.

Task: 259600
Rev-By: alexis