Fixed clipping bug in GL 2 paint engine (visible in arthur demos).
QVectorPath::hints() is not a strict bit field, and thus can not be
anded with RectangleHint. Instead, QVectorPath::shape() should be
directly compared with RectangleHint to check if the vector path is a
rectangle or not. In this case the first four points of a regular
painter path were treated as a rectangle with dire consequences.

Reviewed-by: Tom