Removed over-riding signal declaration from VideoPlayer.
This was causing the stateChanged signal not to propagate up to the Phonon client.  In the case of the mediaplayer demo, this causes the app not to show() the ancestor of the VideoWidget, resulting in the layout not being updated when the clip has been loaded.

This means that the hack to explicitly call show() on this ancestor from the backend has been removed.

Now the layout is correctly updated, but the size of VideoOutput::winId()->Rect() is still incorrect, so the hack to hard-code the screen rectangle passed to CVideoPlayerUtility is left in place.