reverting the definition of WINVER and _WIN32_WINNT in qglobal.h
This just caused too much problems and must be solved another way.
In qfsfileengine_win.cpp we define FSCTL_GET_REPARSE_POINT and all
of the other stuff that's needed for NTFS symlink support, if
its not defined. This is the case if _WIN32_WINNT is less
than 0x0500.
All other changes in this commit are just reversions of commits
that were done for the infamous qglobal.h change.

Discussed with prasanth, tested by pulse.