Improved GLSL precision specifiers in GL 2 engine.
The recommended specifiers are lowp for colors / normal vectors, mediump
for texture coordinates when a limited range is sufficient, and highp
for generic texture coordinates and vertex coordinates / transformation
matrices.

We used to use mediump for texture coordinate in some places, but
since we don't control the texturing scenarios we need to handle the
worst case, which is zooming in on part of a large texture (2048x2048)
with bilinear filtering. To properly handle this case without color
banding mediump is probably not sufficient, so we'll use highp for
texture coordinates.

Reviewed-by: Tom