API improvements for creating shaders from files
It used to be possible to derive the shader type from the file extension,
but this isn't very extensible and doesn't capture the usual extensions.
Change it so that the shader type must be supplied explicitly.

Also add the addShaderFromFile() function to QGLShaderProgram to
provide a convenient short-cut for file-based shader creation.

Reviewed-by: Sarah Smith