Added support for 'make sisx' target in Symbian OS.
Commit f189e00 added support for template PKG file. This commit extends
the template PKG file usage with new make target, i.e. the commit adds
a new make target called sisx, which can be used to generate signed sisx
files.

The sisx target is basically wrapper for calling createpackage.bat,
but it also adds support for default platform/target and environment
variables. Default platform/target feature means that SIS packages are
automatically created for last build target. For example:

>qmake
>make release-armv5
>make sisx                            <- Creates sisx for release-armv5

It is also possible to override the platform and target for which the
SISX is created as follows:

>qmake
>make release-armv5 debug-winscw
>make sisx PLATFORM=ARMV5 TARGET=UREL <- Creates sisx for release-armv5

Since PLATFORM and TARGET are make variables they can also be defined
as an environment variables instead of passing them for make. I.e. the
following is indentical to previous example:

>set PLATFORM=ARMV5
>set TARGET=UREL
>qmake
>make release-armv5 debug-winscw
>make sisx                           <- Creates sisx for release-armv5

The environment variables are also useful if you have your own developer
certificate what you want to use for signing SIS files. For example:

>set CERTIFICATE=mycert.cer
>set KEY=mykey.key
>qmake
>make release-armv5
>make sisx

The above example creates release-armv5 SIS package with custom
certificate and key (key without password). If certificate and key are
not defined, the same logic as in old createpackage.bat will be used i.e.
if RD cert is available in Qt root it will be used, and if not self-signed
cerfificate will be used.

The environment variables supported by 'make sisx' are:
    PLATFORM
    TARGET
    CERTIFICATE
    KEY
    PASSPHRASE

The createpackage.bat can still be called directly but preferred way is
to use new make target directly as examples above demonstrated.

Task: 259037
RevBy: Miikka Heikkinen