Make sure itemAt() reflects the visual order.
This means it should respect the order that was defined with
insertItem() and addItem().
Note that this is not strictly necessary (as now explicitly written in
the docs for QGraphicsLayout::itemAt()), but commit
2ec56d158dc140f68efb45e2e0613f0e4026ddf6 broke the order and for people
that relied on this that commit caused a regression.

In addition, after commit 2ec56d158dc140f68efb45e2e0613f0e4026ddf6 it
was not longer possible to query the "item at visual index". Thus,
instead of adding another function (like QGGL::itemAt(int,int)) we make
sure that itemAt() also returns the "item at visual index".