Reset QTemporaryFile's state after failed open() on Windows
This fixes a regression introduced in 4.5.2 where QTemporaryFile would
no longer attempt to generate a random name after a failed open. Under
certain situations, this led to a non-random file being left behind in
QDir::tempPath when using the fallback implementation of QFile::copy.

Avoid calling QFSFileEngine::setFileName() on a template, so as not to
process it as file name.

By consistently not calling setFileTemplate in the constructor, we also
delay allocation of the fileEngine. Changes made to that function also
keep it from unnecessarily allocating the fileEngine.

Task-number: 260165
Reviewed-by: Thiago Macieira