Rewrite of DirectFB locking mechanism
DirectFB allows you to have a locked subSurface that remains valid while
you paint on the unlocked "parent" surface.

The only limitation is that when accessing the locked memory you might
have to call DirectFB->WaitIdle() in case pending GPU operations aren't
finished.

After this we keep the locked surface around at all times (from the
first time it's requested) until the surface dies. Previous calls to
lock() will just call WaitIdle if necessary and previous calls to unlock
now just mark the surface as dirty and in need of a WaitIdle if someone
needs to access its pixel data.

Reviewed-by: Donald Carr <donald.carr@nokia.com>