Changed createpackage parameters to be more consistent with make targets.
Earlier createpackage required two separate parameters to define the
platform and target. For example
    > createpackage fluidlauncher_template.pkg release armv5

Now it is changed to:
    > createpackage fluidlauncher_template.pkg release-armv5

This is consistent to make target what is used to do the building i.e.:
    > make release-armv5

The change also affected environment variables supported by 'make sisx'
target. QT_SISX_PLATFORM is not any more supported, and the info is
merged QT_SISX_TARGET variable.

Reviewed-by: Miikka Heikkinen