Enable Q_OBJECT usage across DLL boundary on Symbian OS
Symbian loader does not allow DLL import relocations in the data
section. To workaround this, added an accessor function which can be
used to get the class static metadata from the base class even if it is
in another DLL.

Changing the implementation for all platforms would cause a binary
break, so it is only used for Symbian. The Q_NO_DATA_RELOCATION macro
is used to configure this feature instead of Q_OS_SYMBIAN, because
it is possible another platform may need the same fix and Symbian may
eventually fix their DLL loader to support this feature.

Task-number: 258893
Reviewed-by: Kent Hansen