Fixed crash when sharing OpenGL contexts in the GL2 paint engine.
Shader objects had a pointer to the context they were originally
created in. If the context was destroyed, the shader would (on Windows)
dereference an invalid pointer and cause the program to crash. I
replaced the context pointer with a pointer to the context group. I
also added checks in debug mode to make sure the context associated
with the shader shares resources with the current context.

Reviewed-by: Tom