Using QFocusFrame to visualize the focus.
When using the keypad navigation, the focus state of a widget is
supposed to be well visible. So it makes much sense to use QFocusFrame,
which is able to paint a focus marker _around_ the widget.

This change should fix the issue that with some S60 themes, the focus
was invisible.

Also, a lot of code is removed by this change.

Task-number: 256470
Reviewed-by: Sami Merila

	modified:   src/gui/styles/qs60style.cpp
	modified:   src/gui/styles/qs60style.h
	modified:   src/gui/styles/qs60style_p.h