Removed hard-coded sqlite3 deployment statement - used PKG conditions.
Before this change Qt did not install to device which had sqlite3
already in ROM. Since we now install the sqlite3 only if it is not on
any drive, the installation works also with preinstalled sqlite.

Even better appraoch would be to use embedded SIS file for sqlite,
since that would allow sqlite3 upgrade without upgrading the whole Qt.
However we yet don't have such SIS, but once we get one from Symbian
this needs to be further changed.

Task-number: 258858
Reviewed-by: Miikka Heikkinen