Fixed rendering bug in blurpicker example with -graphicssystem opengl
Made the GL blur filter code slightly less hacky by not reusing the same
paint engine for rendering both to the offscreen FBO and to the actual
target device. This should make the code less reliant on paint engine
implementation details and thus more robust with regards to changes in
the paint engine.

Task-number: 260402
Reviewed-by: Trond