Speed-up floating-point decoding for SVG parsing.
Instead of comparing the character to '0' and '9', use bit fiddling
to detect that the character is a digit between '0' and '9'.

Loading tiger.svg (tests/benchmarks/qsvgrenderer) is now 10% faster,
going down from 85.3 millions instructions to 77.2 millions.
Mostly this is due 46% speed-up in parseNumbersList() function,
from 26.9 millions instructions to just 18.4 millions.

Reviewed-by: Kim