Disable event compression when the tablet events are accepted
Since Qt 4.5, all tablet events are compressed not to overload the
widgets with the corresponding mouse event (a mouse event is generated
if the tablet event is not accepted).

This behavior reduce the precision when drawing on a widget that use
the tablet events. All tablet events should be sent to the widget that
are accepting the tablet event.

With this patch, the tablet event are filtered only if the widget
ignore the first tablet event. The mouse events are compressed.

There is two special cases for the filtering:

First, if a tablet event is for another widget than the one
ignoring the tablet, this event should not be filtered.

Second, if there is a mouse press event, the mouse move event should be
sent to the widget that received the mouse press event.

Helped-by: Pierre Rossi
Reviewed-by: Thomas Zander
Reviewed-by: Bradley T. Hughes