Make sure the BSP is updated when a graphics effect changes bounding rect.
In the case of applying an effect to a QGraphicsItem, we have to notify
the scene's BSP that the item's bounding rect has changed. We do this
by calling prepareGeometryChange().

In the case of QWidget, it's sub-optimal that we update its parent, but
there's no other way to solve it at the moment.