Make QGLFramebufferObject work again using new QGLPaintDevice API
This patch also refactors QGL2PaintEngineEx::ensureActive() and the
logic which handles multiple paint engines rendering to the same
QGLContext. In a nut-shell:

* QGLPaintDevice::beginPaint() stores the currently bound FBO
* QGLPaintDevice::ensureActiveTarget() makes sure that GL rendering will
  end up in the paint device (I.e. the right context is current and the
  right FBO is bound). If a different context or FBO was bound, it is
  _not_ remembered.
* QGLPaintDevice::endPaint() restores whatever FBO was bound when
  beginPaint() was called.

This logic allows interleaved painter rendering to multiple FBOs and
contexts to work as expected. It also allows a stacked begin/end to work
properly when it's mixed with native GL rendering (as far as current
render target is concerened. GL state clobbering is obviously a
different topic).

QGLPaintDevice::context() also had to be made virtual as there's no good
place to call setContext. This might be possible to change in the future
though.

Finally, to make this work, QGLFramebufferObjectPrivate had to be moved
into it's own private header.