Speed-up parsing of SVG path data.
Instead of using operations that shuffle the array of numbers, just
use pointer to iterate the numbers. This reduced the amount of
memory operations during the parsing.

In addition, parse the numbers to QVarLengthArray instead of QVector.
This works well because usually a path element is typically followed by
a short list of numbers.

Loading tiger.svg (tests/benchmarks/qsvgrenderer) is now 8% faster,
mostly due to the time spent in parsePathDataFast is reduced from
26.1 millions instructions to just 20.5 millions (27% speed-up).

Reviewed-by: Kim