Make QTreeModel::ensureSorted() stable sort for items
Id you have numerous items with the same value in the sort column,
whenever you update one of them, they'll be placed at the head of
the list instead of staying in place.

For example, assume you have items
       a b(1) b(2) b(3) b(4) c
(where all the b have the same value in the sort column)
If you now emitDataChanged from b(3), ensureSorted() will be called.
It will place b(3) in a list, and stable sort the list. It's just
the one item since there was only one item updated.

It than takes each item in the list, removes it's place from the
"full" list of items, then reinserts it at the earliest point (using qLowerBound).

End result:
       a b(3) b(1) b(2) b(4) c

If you update all the items in the list (doing emitDataChanged()
for each), this has the effect of reversing all the items with
identical sort key.

This patch checks if the old row is within the lower and upper
bound of where the item might go, and if it is, simply reinserts
 it in its old place.

Reviewed-by: Olivier Goffart
Merge-Request: 1393