Added macros to QtGui for controlling whether widgets are window-owning
There are two macros defined in src/gui/kernel/window_owning_control.h:

If this is defined, then calling QWidget::winId causes a native window to be created for the associated control

If QWIDGET_OWNS_WINDOW is defined, defining this macro means that the parent window for the newly created window is the parent widget's winId.  This in turn means that calling QWidget::winId causes native windows to be created for all ancestors of the target widget.
If this macro is undefined, the parent window is the parent widget's effectiveWinId.