Fix test failure & optimize matrix operations in QGraphicsItem.
This fixes failures in tst_QGraphicsItem::setTransformProperties().

Change 9e8ff32d introduced QMatrix4x4 as an internal matrix for
QGraphicsItem. Problem is, QMatrix4x4 is float-based whereas QTransform
is double-based. This change readds the use of QTransform in the case
where there are no QGraphicsTransforms in the list. This by itself also
makes this common case a bit faster.

The workaround is moot if somebody adds any QGraphicsTransform,
including one that doesn't do anything (like rotate by 0 degrees). So
we might have to find a better fix.

Reviewed-by: Olivier