Improvements to usage of QMatrix4x4 in QGraphicsItem.
This could have been amended to 464ee98eee5ce160b497aec6a1163422980dd797,
the idea is that we use QMatrix4x4 only for the QGraphicsTransform
instances and QTransform for the rest. The problem with the last submit
was that adding any (unrelated) QGraphicsTransform would cause the
QGraphicsItem::rotation properties (and friends) to lose precision.

Reviewed-by: Olivier
Reviewed-by: gabi