Rendering artifacts when installing an effect on HasNoContents items.
QGraphicsItem::HasNoContents is documented to not paint anything when
the flag is set on an item, so all the update requests are ignored. However,
we cannot ignore update requests on such items when an effect is
installed on them, because when the effects changes parameters it calls
update on the item. We still don't paint the item, but we processes the
update request such that its children can be painted properly.

Reviewed-by: Andreas