Make itemviews use gradient selection on Vista by default
We postponed this in 4.3 when the vista style was introduced
since a lot of microsoft apps still used the legacy style
and microsoft did not enable it by default.

It has been reported as bugs a few times and it seems increasingly
odd that we are not using the more modern look and feel when we
can. Instead of only enabling it for QTreeView, we now
only disable it for QTableView as the look still cannot be used
to span vertical cells.

Task-number: QTBUG-3746
Reviewed-by: prasanth