Modify QMatrix4x4 and QQuaternion to use qreal internally
Some concerns were expressed about the float precision of QMatrix4x4,
which this change addresses by using qreal instead.

The QVector2D/3D/4D classes still use float internally, so that they
can be used directly in large arrays of vertex values to be uploaded
to an OpenGL server.

QQuaternion is a client-side class, and it should produce rotations
that are consistent with QMatrix4x4.  So its precision was changed too.

A consequence of this change is that the following no longer works
in a portable fashion:

    QMatrix4x4 mat;
    ...
    glLoadMatrixf(mat.constData());

The caller must now repack the argument to convert from qreal to GLfloat.

Reviewed-by: Michael Goddard
Reviewed-by: Andreas